renaming back plugin/src/main/java/br to plugin/src/br
