+++ br.usp.each.saeg.jaguar.plugin/src/br/usp/each/saeg/jaguar/plugin/JaguarConfigTabGroup.java
@@ -1,0 +1,9 @@
+package br.usp.each.saeg.jaguar.plugin;
+
+import org.eclipse.pde.ui.launcher.JUnitTabGroup;
+
+public class JaguarConfigTabGroup extends JUnitTabGroup {
+
+
+
+}
