+++ br.usp.each.saeg.jaguar.plugin/src/br/usp/each/saeg/jaguar/plugin/LaunchConfigurationsMessages.properties
