--- br.usp.each.saeg.jaguar.plugin/src/main/java/br/usp/each/saeg/jaguar/plugin/JaguarConfigTabGroup.java
@@ -1,9 +1,0 @@
-package br.usp.each.saeg.jaguar.plugin;
-
-import org.eclipse.pde.ui.launcher.JUnitTabGroup;
-
-public class JaguarConfigTabGroup extends JUnitTabGroup {
-
-
-
-}
